refactor: refactor result and aggregators in utils (#99)
* refactor: refactor result and aggregators in utils

* refactor: refactor test of results in utility package
