+++ src/test/scala/dsl/util/ScoobyTest.scala
@@ -3,7 +3,7 @@ package dsl.util
 
 import core.scooby
 import core.scooby.{Configuration, SingleExporting}
-import core.scraper.Result
+import org.unibo.scooby.utility.result.Result
 import dsl.DSL.ConfigurationWrapper
 import dsl.ScoobyEmbeddable
 import utility.document.{CrawlDocument, ScrapeDocument}
